requirements.txt
pandas
dspy
google-genai
nltk
scikit-learn
sentence-transformers
torch